f32 to f64

+++ sdl2-mandelbrot/src/mandelbrot.rs
@@ -9,16 +9,16 @@ use sdl2::rect::Rect;
 use sdl2::image::LoadTexture;
 use utils;
 
-const R1: Complex<f32> = Complex {re: -2.0, im: -1.0};
-const R2: Complex<f32> = Complex {re: 1.0, im: 1.0};
+const R1: Complex<f64> = Complex {re: -2.0, im: -1.0};
+const R2: Complex<f64> = Complex {re: 1.0, im: 1.0};
 const ITERATIONS: u32 = 100;
 
 pub struct Mandelbrot {
     creator: TextureCreator<WindowContext>,
     texture: Option<Texture>,
     pub render_time: Option<Duration>,
-    r1: Complex<f32>,
-    r2: Complex<f32>,
+    r1: Complex<f64>,
+    r2: Complex<f64>,
 }
 
 impl Mandelbrot {
@@ -109,18 +109,18 @@ impl Mandelbrot {
 fn pixel_to_point(x: usize,
                   y: usize,
                   bounds: (usize, usize),
-                  top_left: &Complex<f32>,
-                  bottom_right: &Complex<f32>) -> Complex<f32> {
+                  top_left: &Complex<f64>,
+                  bottom_right: &Complex<f64>) -> Complex<f64> {
 
     let (width, height) = (bottom_right.re - top_left.re,
                            top_left.im - bottom_right.im);
     Complex {
-        re: top_left.re + x as f32 * width / bounds.0 as f32,
-        im: top_left.im - y as f32 * height / bounds.1 as f32,
+        re: top_left.re + x as f64 * width / bounds.0 as f64,
+        im: top_left.im - y as f64 * height / bounds.1 as f64,
     }
 }
 
-pub fn escape_time(c: Complex<f32>, limit: u32) -> Option<u32> {
+pub fn escape_time(c: Complex<f64>, limit: u32) -> Option<u32> {
     let mut z = Complex {re: 0.0, im: 0.0};
     for i in 0..limit {
         z = z*z + c;
